refactor(datagrid): split 2,251-line file into 14 focused modules
Modules created:
- core.rs (489 lines): struct definition, builders, public API
- render.rs (419 lines): rendering logic
- editing.rs (258 lines): cell editing
- tree.rs (221 lines): tree grid mode
- reorder.rs (135 lines): column reorder
- navigation.rs (135 lines): selection/navigation
- export.rs (122 lines): export functionality
- mouse.rs (113 lines): mouse handling
- resize.rs (93 lines): column resize
- filter.rs (70 lines): sort/filter
- footer.rs (49 lines): aggregation footer
- width.rs (47 lines): width calculation
- freeze.rs (33 lines): column freeze

mod.rs reduced from 2,251 to 44 lines (-98%)
All 203 tests pass
